pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evo-agent"
version = "0.3.0"
description = "Evolutionary Prompt Manager: context-aware, feedback-driven AI code assistant with UI and multithread scanning."
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1.7",
  "tiktoken>=0.7.0",
  "PyYAML>=6.0.2",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.6",
  "jinja2>=3.1.4"
]

[project.optional-dependencies]
llm = ["openai>=1.51.0", "anthropic>=0.39.0"]
vector = ["chromadb>=0.5.15"]

[project.scripts]
evo-agent = "evo_agent.cli:main"
evo-agent-ui = "evo_agent.server:run"

[tool.setuptools.packages.find]
where = ["src"]